[codex][refactor] 引入运行时模块分层骨架
